+++ framework/core/src/test/java/com/tenio/core/network/entity/session/implement/SessionImplTest.java
@@ -39,6 +39,6 @@ class SessionImplTest {
     assertFalse(actualNewInstanceResult.isAssociatedToPlayer(Session.AssociatedState.DONE));
     assertFalse(actualNewInstanceResult.isActivated());
     assertEquals(TransportType.UNKNOWN, actualNewInstanceResult.getTransportType());
-    assertNull(actualNewInstanceResult.fetchPacketQueue());
+    assertNull(actualNewInstanceResult.fetchOutboundQueue());
   }
 }